Move msw mocks to `/domain`.
