+++ coral/src/domain/api-mocks/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from "msw";
-import { handlers } from "src/services/http-client/mocks/handlers";
+import { handlers } from "src/domain/api-mocks/handlers";
 
 // This configures a Service Worker with the given request handlers.
 export const worker = setupWorker(...handlers);
