+++ coral/src/domain/api-mocks/server.ts
@@ -1,4 +1,4 @@
 import { setupServer } from "msw/node";
-import { handlers } from "src/services/http-client/mocks/handlers";
+import { handlers } from "src/domain/api-mocks/handlers";
 
 export const server = setupServer(...handlers);
